[STD-26] Tabs - Starting tabs rework - Moving systems and creating path alias
