+++ webview/lib/ProjectionEngine.ts
@@ -1,18 +1,18 @@
 import IInjectable from "@lib/IInjectable";
-import ToasterService from "../window/services/ToasterService";
-import ViewportActionService from "../window/services/ViewportActionService";
+import ToasterService from "@services/ToasterService";
+import ViewportActionService from "@services/ViewportActionService";
 import ContentBrowserStore from "./stores/ContentBrowserStore";
-import DragDropService from "../window/services/DragDropService";
-import EntityNamingService from "../window/services/EntityNamingService";
-import EntityUpdateService from "../window/services/EntityUpdateService";
-import LevelService from "../window/services/LevelService";
-import ExecutionService from "../window/services/ExecutionService";
+import DragDropService from "@services/DragDropService";
+import EntityNamingService from "@services/EntityNamingService";
+import EntityUpdateService from "@services/EntityUpdateService";
+import LevelService from "@services/LevelService";
+import ExecutionService from "@services/ExecutionService";
 import EngineStore from "./stores/EngineStore";
 import SettingsStore from "./stores/SettingsStore";
 import TabsStore from "./stores/TabsStore";
 import Engine from "@engine-core/Engine";
-import EntityHierarchyService from "../window/services/EntityHierarchyService";
-import ContextMenuService from "../window/services/ContextMenuService";
+import EntityHierarchyService from "@services/EntityHierarchyService";
+import ContextMenuService from "@services/ContextMenuService";
 import EntitySelectionStore from "./stores/EntitySelectionStore";
 import {Inject} from "@lib/Injection";
