+++ webview/services/LevelService.ts
@@ -1,17 +1,17 @@
 import FileSystemUtil from "@lib/FileSystemUtil"
 import Engine from "@engine-core/Engine"
-import EditorFSUtil from "../editor/util/EditorFSUtil"
+import EditorFSUtil from "../window/editor/util/EditorFSUtil"
 import EntitySelectionStore from "@lib/stores/EntitySelectionStore"
 import CameraAPI from "@engine-core/lib/utils/CameraAPI"
-import CameraTracker from "../../engine/tools/utils/CameraTracker"
+import CameraTracker from "@engine-tools/utils/CameraTracker"
 import serializeStructure from "@engine-core/utils/serialize-structure"
 import QueryAPI from "@engine-core/lib/utils/QueryAPI"
 import PickingAPI from "@engine-core/lib/utils/PickingAPI"
-import AXIS from "../../engine/tools/static/AXIS"
+import AXIS from "@engine-tools/static/AXIS"
 import LocalizationEN from "@enums/LocalizationEN"
 import FileTypes from "@enums/FileTypes"
-import EditorUtil from "../editor/util/EditorUtil"
-import TabsStoreUtil from "../editor/util/TabsStoreUtil"
+import EditorUtil from "../window/editor/util/EditorUtil"
+import TabsStoreUtil from "../window/editor/util/TabsStoreUtil"
 import {Inject, Injectable} from "@lib/Injection";
 import VisualsStore from "@lib/stores/VisualsStore";
 import TabsStore from "@lib/stores/TabsStore";
