+++ webview/window/editor/templates/get-viewport-context.ts
@@ -1,5 +1,5 @@
 import getViewportHotkeys from "./get-viewport-hotkeys"
-import EntityFactoryService from "../../services/EntityFactoryService"
+import EntityFactoryService from "@services/EntityFactoryService"
 import QueryAPI from "@engine-core/lib/utils/QueryAPI"
 import getEntityCreationOptions from "./get-entity-creation-options"
 import ContextMenuOption from "@lib/context-menu/templates/ContextMenuOptions"
