+++ webview/window/editor/templates/get-viewport-hotkeys.ts
@@ -1,11 +1,11 @@
 import GIZMOS from "@enums/Gizmos"
 import QueryAPI from "@engine-core/lib/utils/QueryAPI"
 import GizmoTransformationType from "@enums/GizmoTransformationType"
-import EntityFactoryService from "../../services/EntityFactoryService"
+import EntityFactoryService from "@services/EntityFactoryService"
 import CAMERA_ROTATIONS from "../../../engine/tools/static/CAMERA_ROTATIONS"
 import CameraTracker from "../../../engine/tools/utils/CameraTracker"
 import ContextMenuOption from "@lib/context-menu/templates/ContextMenuOptions"
-import EngineStateService from "../../services/EngineStateService"
+import EngineStateService from "@services/EngineStateService"
 import LocalizationEN from "@enums/LocalizationEN"
 import EditorUtil from "../util/EditorUtil"
 import EntitySelectionStore from "@lib/stores/EntitySelectionStore";
