+++ webview/window/editor/util/HierarchyUtil.ts
@@ -1,8 +1,7 @@
 import HierarchyToRenderElement from "../views/hierarchy/template/ToRenderElement"
 import LocalizationEN from "@enums/LocalizationEN"
 import Entity from "@engine-core/instances/Entity"
-import Engine from "@engine-core/Engine"
-import EngineStateService from "../../services/EngineStateService"
+import EngineStateService from "@services/EngineStateService"
 import EditorUtil from "./EditorUtil"
 import HotKeysController from "@lib/HotKeysController";
 import getViewportHotkeys from "../templates/get-viewport-hotkeys";
