+++ webview/window/editor/views/hierarchy/components/Header.svelte
@@ -6,7 +6,7 @@
     import Input from "@lib/components/input/Input.svelte"
     import getDropdownHeaderStyles from "@lib/components/dropdown/utils/get-dropdown-header-styles"
     import NATIVE_COMPONENTS from "../../inspector/static/NATIVE_COMPONENTS"
-    import EntityFactoryService from "../../../../services/EntityFactoryService"
+    import EntityFactoryService from "@services/EntityFactoryService"
     import LocalizationEN from "@enums/LocalizationEN"
     import EmptyIcon from "@lib/components/icon/EmptyIcon.svelte"
     import ProjectionEngine from "@lib/ProjectionEngine";
