+++ webview/window/editor/views/preferences/components/content/ContentGroup.svelte
