+++ webview/window/editor/views/preferences/static/PREFERENCES_TYPES.ts
